add assets for icons and organized components to appropriated forlders
