+++ web-app/.eslintrc.yml
@@ -1,6 +1,7 @@
 env:
   browser: true
   es2021: true
+  node: true
 extends:
   - eslint:recommended
   - plugin:react/recommended
