+++ web-app/src/atoms/LoginButton.js
@@ -1,0 +1,11 @@
+import React from "react";
+import { useAuth0 } from "@auth0/auth0-react";
+
+function LoginButton() {
+
+	const { loginWithRedirect, isAuthenticated } = useAuth0();
+    
+	return !isAuthenticated && <button onClick={() => loginWithRedirect()}>Log In</button>;
+}
+  
+export default LoginButton;
