+++ web-app/src/atoms/LogoutButton.js
@@ -3,9 +3,9 @@ import { useAuth0 } from "@auth0/auth0-react";
 
 function LogoutButton() {
 
-    const { logout, isAuthenticated } = useAuth0();
+	const { logout, isAuthenticated } = useAuth0();
     
-    return isAuthenticated && <button onClick={() => logout()}>Log Out</button>;
-  };
+	return isAuthenticated && <button onClick={() => logout()}>Log Out</button>;
+}
   
 export default LogoutButton;
