+++ web-app/src/atoms/icons.js
@@ -1,0 +1,34 @@
+import React from "react";
+
+
+export function ChainIcon() {
+	return <img alt="link" src="/assets/images/chain.png" className="icon"/>;
+}
+
+export function CrossIcon() {
+	return <img alt="cross" src="/assets/images/delete.png" className="icon"/>;
+}
+
+export function DragVerticalIcon() {
+	return <img alt="drag vertical" src="/assets/images/drag_vertical.png" className="icon"/>;
+}
+
+export function GearIcon() {
+	return <img alt="gear" src="/assets/images/gear.png" className="icon"/>;
+}
+
+export function LogoIcon() {
+	return <img alt="bookmark" src="/assets/images/logo.png" className="icon"/>;
+}
+
+export function PencilIcon() {
+	return <img alt="edit" src="/assets/images/pencil.png" className="icon"/>;
+}
+
+export function ShareIcon() {
+	return <img alt="share" src="/assets/images/share.png" className="icon"/>;
+}
+
+export function TrashCanIcon() {
+	return <img alt="trash can" src="/assets/images/trash_can.png" className="icon"/>;
+}
