--- web-app/src/components/TextBox.js
@@ -1,14 +1,0 @@
-import React from "react";
-import './TextBox.css';
-class TextBox extends React.Component {
-    render() {
-        return (
-            <div className= "text-box">
-                <b>My Inspiration</b>
-                <div>It does not matter how fast you go as long as you do not stop.</div>
-            </div>
-        )
-    }
-}
-
-export default TextBox;
