+++ web-app/src/index.js
@@ -1,28 +1,28 @@
 // built-in
-import React from 'react';
-import ReactDOM from 'react-dom';
+import React from "react";
+import ReactDOM from "react-dom";
 
 // local
-import './index.css';
-import App from './App';
+import "./index.css";
+import App from "./App";
 
 // third-party
 import { Auth0Provider } from "@auth0/auth0-react";
 
 const domain = process.env.REACT_APP_AUTH0_DOMAIN;
-const clientId= process.env.REACT_APP_AUTH0_CLIENT_ID
+const clientId= process.env.REACT_APP_AUTH0_CLIENT_ID;
 
 ReactDOM.render(
-  <React.StrictMode>
-    <Auth0Provider
-      domain={domain}
-      clientId={clientId}
-      redirectUri={window.location.origin}
-    >
-      <App />
-    </Auth0Provider>
-  </React.StrictMode>,
-  document.getElementById("root")
+	<React.StrictMode>
+		<Auth0Provider
+			domain={domain}
+			clientId={clientId}
+			redirectUri={window.location.origin}
+		>
+			<App />
+		</Auth0Provider>
+	</React.StrictMode>,
+	document.getElementById("root")
 );
